LayoutTests/media/modern-media-controls/start-button/start-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>StartButton</code> class.");

const startButton = new StartButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("startButton.element.localName", "button");
shouldBeTrue('startButton.element.classList.contains("start")');
shouldBe("startButton.element.childElementCount", "2");
shouldBeEqualToString("startButton.element.firstElementChild.localName", "div");
shouldBeEqualToString("startButton.element.firstElementChild.className", "background");
shouldBeEqualToString("startButton.element.lastElementChild.localName", "img");
shouldBeTrue("startButton.element.lastElementChild.src.includes('macOS/start@')");

const iOSStartButton = new StartButton({ layoutTraits: LayoutTraits.iOS });
shouldBeTrue("iOSStartButton.element.lastElementChild.src.includes('iOS/start@')");


</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
